Add files via upload
This was a college task to estimate the Generalized Renewal Processes (GRP) parameters on machine failure times using the maximum likelihood method. I opted to use the Monte Carlo simulation as a numeric method to get pretty reliable values for those, and later used the same numeric method to simulate a scenario where i could calculate the average error compared to real data.